% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_qqplots.R
\name{generate_qqplots}
\alias{generate_qqplots}
\title{Generate QQ Plots}
\usage{
generate_qqplots(dataset, dataset_subvector = NULL)
}
\arguments{
\item{dataset}{The dataset to generate the QQ plots for.}

\item{dataset_subvector}{(Optional) The subvector of variables to generate QQ plots for.}
}
\description{
This function generates QQ plots for a specified subvector of a dataset. If no
subvector is specified, the function iterates over all variables in the dataset,
creating QQ plots for numeric variables and ignoring non-numeric variables.
}
\details{
Generate QQ plots for a subvector of a dataset
}
\examples{
generate_qqplots(mtcars, c("mpg", "cyl", "disp"))

}
