% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_residual_plots.R
\name{generate_residual_plots}
\alias{generate_residual_plots}
\title{Generate Residual Plots}
\usage{
generate_residual_plots(dataset, target_variable, dataset_subvector = NULL)
}
\arguments{
\item{dataset}{The dataset to generate the residual plots for.}

\item{target_variable}{The target variable (dependent variable) for the residual plots.}

\item{dataset_subvector}{(Optional) The subvector of variables to generate residual plots for.}
}
\description{
This function generates residual plots for a specified target variable and subvector of
variables in a dataset. If no subvector is specified, the function iterates over all
variables in the dataset, creating residual plots for numeric variables and ignoring
non-numeric variables.
}
\details{
Generate residual plots for a target variable & subvector of a dataset
}
\examples{
generate_residual_plots(mtcars, "mpg", c("cyl", "disp"))

}
