% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_scatterplots.R
\name{generate_scatterplots}
\alias{generate_scatterplots}
\title{Generate Scatterplots with Regression Lines}
\usage{
generate_scatterplots(dataset, target_variable, dataset_subvector = NULL)
}
\arguments{
\item{dataset}{The dataset to generate the scatterplots for.}

\item{target_variable}{The target variable (dependent variable) for the scatterplots.}

\item{dataset_subvector}{(Optional) The subvector of variables to generate scatterplots for.}
}
\description{
This function generates scatterplots with regression lines for a specified target variable
and subvector of variables in a dataset. If no subvector is specified, the function iterates
over all variables in the dataset, creating scatterplots for numeric variables and ignoring
non-numeric variables.
}
\details{
Generate scatterplots for a target variable & subvector of a dataset
}
\examples{
generate_scatterplots(mtcars, "mpg", c("cyl", "disp"))

}
